requirements.txt
flask==3.0.2
git+https://github.com/openai/openai-python.git
python-dotenv==1.0.1
PyPDF2==3.0.1
python-docx==1.0.1
gunicorn==21.2.0
whitenoise==6.6.0
requests==2.31.0
werkzeug==3.0.1 